backend/core_scraper.py
import os
import time
from docx import Document as DocxDocument
import PyPDF2
import pandas as pd
from pptx import Presentation
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from bs4 import BeautifulSoup
import requests

def mcp_analyze(text):
    # Replace with real MCP API / SDK
    return {"summary": text[:200], "word_count": len(text.split())}

def extract_docx(path):
    doc = DocxDocument(path)
    return "\n".join([p.text for p in doc.paragraphs])

def extract_pdf(path):
    text = ""
    with open(path, "rb") as f:
        reader = PyPDF2.PdfReader(f)
        for page in reader.pages:
            text += page.extract_text() or ""
    return text

def extract_csv(path):
    df = pd.read_csv(path)
    return df.to_string()

def extract_xlsx(path):
    df = pd.read_excel(path)
    return df.to_string()

def extract_pptx(path):
    prs = Presentation(path)
    text = ""
    for slide in prs.slides:
        for shape in slide.shapes:
            if hasattr(shape, "text"):
                text += shape.text + "\n"
    return text

def scrape_site(url, use_js, doc_types, out_dir, job_id):
    # Recursive scraping logic + document download/extraction
    # For brevity, imagine the same logic as prior messages, but with:
    # - Download docs
    # - Extract content
    # - MCP analysis
    # - Save txt/docx files
    # - Update job status/progress
    pass  # Full implementation omitted for brevity, but would be based on previous code