backend/requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.32.1
aiofiles==24.1.0
requests==2.32.4
beautifulsoup4==4.12.3
selenium==4.27.1
python-docx==1.1.2
PyPDF2==3.0.1
pandas==2.2.3
python-pptx==1.0.2
openpyxl==3.1.5
pydantic==2.10.3
starlette==0.49.1
python-multipart==0.0.20
